libwasi: move poll into a separate file
